+++ packages/core/src/storage/state.js
@@ -29,6 +29,7 @@ export default class {
 
         // Private stores
         this._queue = writable(new Map());  // debounce queue
+        this._auxiliary = writable([]);  // auxiliary components
         this._currentRange = currentRange(this);
         this._activeRange = activeRange(this);
         this._fetchedRange = writable({start: undefined, end: undefined});
@@ -46,6 +47,8 @@ export default class {
         this._view = view2(this);
         this._viewComponent = writable(undefined);
         this._viewClass = writable(undefined);
+        // Resources
+        this._resBgColor = writable(noop);
         // Interaction
         this._interaction = writable({});
         this._iEvents = writable([null, null]);  // interaction events: [drag/resize, pointer]